# api/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Self-hosted media server for movies, TV shows, and music",
        "title": "ViewRA Media Server API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "ViewRA Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/tv/episodes/{id}": {
            "get": {
                "description": "Returns details of a specific TV episode including all metadata",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tv"
                ],
                "summary": "Get a TV episode by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Episode ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_viewra_viewra_internal_application_tv.TVEpisodeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/tv/search": {
            "get": {
                "description": "Searches for TV shows and episodes by title in a specific library",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tv"
                ],
                "summary": "Search TV shows and episodes",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Library ID to search in",
                        "name": "library_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Search query (show title or episode title)",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_viewra_viewra_internal_application_tv.ListTVEpisodesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/tv/shows": {
            "get": {
                "description": "Returns a list of all TV shows in a library with episode and season counts and optional pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tv"
                ],
                "summary": "List TV shows",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Library ID to filter shows",
                        "name": "library_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page (default: 50, max: 200)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items to skip (default: 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_viewra_viewra_internal_application_tv.ListTVShowsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/tv/shows/{id}": {
            "get": {
                "description": "Returns details of a specific TV show including season and episode counts",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tv"
                ],
                "summary": "Get a TV show by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Show ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_viewra_viewra_internal_application_tv.TVShowDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/tv/shows/{id}/episodes": {
            "get": {
                "description": "Returns all episodes for a specific TV show using its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tv"
                ],
                "summary": "List episodes for a TV show by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Show ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_viewra_viewra_internal_application_tv.ListTVEpisodesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/tv/shows/{showTitle}/episodes": {
            "get": {
                "description": "Returns all episodes for a specific TV show",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tv"
                ],
                "summary": "List episodes for a TV show",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Show title",
                        "name": "showTitle",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Library ID",
                        "name": "library_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_viewra_viewra_internal_application_tv.ListTVEpisodesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_viewra_viewra_internal_application_tv.ListTVEpisodesResponse": {
            "type": "object",
            "properties": {
                "episodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_viewra_viewra_internal_application_tv.TVEpisodeResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_viewra_viewra_internal_domain_common.PaginationMetadata"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "github_com_viewra_viewra_internal_application_tv.ListTVShowsResponse": {
            "type": "object",
            "properties": {
                "pagination": {
                    "$ref": "#/definitions/github_com_viewra_viewra_internal_domain_common.PaginationMetadata"
                },
                "shows": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_viewra_viewra_internal_application_tv.TVShowSummary"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "github_com_viewra_viewra_internal_application_tv.TVEpisodeResponse": {
            "type": "object",
            "properties": {
                "air_date": {
                    "type": "string"
                },
                "audio_codec": {
                    "type": "string"
                },
                "bitrate": {
                    "type": "integer"
                },
                "container_format": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration": {
                    "description": "in seconds",
                    "type": "integer"
                },
                "episode": {
                    "type": "integer"
                },
                "episode_title": {
                    "description": "actual episode name",
                    "type": "string"
                },
                "file_path": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "frame_rate": {
                    "type": "number"
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "description": "Base Media fields",
                    "type": "integer"
                },
                "imdb_id": {
                    "type": "string"
                },
                "is_extra": {
                    "type": "boolean"
                },
                "library_id": {
                    "type": "integer"
                },
                "season": {
                    "type": "integer"
                },
                "season_id": {
                    "type": "integer"
                },
                "show_title": {
                    "description": "TV Episode-specific fields",
                    "type": "string"
                },
                "title": {
                    "description": "filename",
                    "type": "string"
                },
                "tvdb_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "video_codec": {
                    "type": "string"
                },
                "width": {
                    "description": "Technical metadata",
                    "type": "integer"
                }
            }
        },
        "github_com_viewra_viewra_internal_application_tv.TVShowDetailResponse": {
            "type": "object",
            "properties": {
                "episode_count": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "library_id": {
                    "type": "integer"
                },
                "season_count": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "github_com_viewra_viewra_internal_application_tv.TVShowSummary": {
            "type": "object",
            "required": [
                "id",
                "library_id",
                "title"
            ],
            "properties": {
                "episode_count": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "library_id": {
                    "type": "integer"
                },
                "season_count": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "github_com_viewra_viewra_internal_domain_common.PaginationMetadata": {
            "type": "object",
            "properties": {
                "hasMore": {
                    "description": "Whether there are more items available",
                    "type": "boolean"
                },
                "limit": {
                    "description": "Number of items per page",
                    "type": "integer"
                },
                "offset": {
                    "description": "Number of items skipped",
                    "type": "integer"
                },
                "total": {
                    "description": "Total number of items across all pages",
                    "type": "integer"
                }
            }
        },
...
        "internal_api_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}